requirements.txt
base58==2.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
Flask==1.1.2
idna==2.9
ipfshttpclient==0.4.12
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
multiaddr==0.0.9
mysql==0.0.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
netaddr==0.7.19
numpy==1.18.2
pandas==1.0.3
protobuf==3.6.1
psutil==5.7.0
python-dateutil==2.8.1
python-gnupg==0.4.5
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
varint==1.0.2
Werkzeug==1.0.1
